[ADD]分离注入函数，创建 v3 框架
